+++ NEWS.md
@@ -1,3 +1,5 @@
+# BFS (development version)
+
 # BFS 0.4.0
 
 * BREAKING CHANGE: `clean_names` of `bfs_get_dataset()` is now FALSE by default.