+++ NEWS.md
@@ -2,9 +2,9 @@
 
 # BFS 0.4.0
 
-* BREAKING CHANGE: `clean_names` of `bfs_get_dataset()` is now FALSE by default.
-* `bfs_get_dataset()` uses BFS API v1.
-* `bfs_get_metadata()` removed because not working. Use only `bfs_get_catalog_data()` for now.
+* BREAKING CHANGE: `clean_names` of `bfs_get_data()` is now FALSE by default.
+* `bfs_get_data()` uses BFS API v1.
+* `bfs_get_metadata()` deprecated. Please use only `bfs_get_catalog_data()`.
 * add `bfs_get_catalog_tables()`.
 * add `bfs_get_catalog()`.
 * pxweb and jsonlite dependency added. dplyr dependency removed.